fix(codegen): proper array construction
